requirements.txt
torchsummary~=1.5.1
numpy~=1.21.4
sklearn~=0.0
scikit-learn~=0.24.0
pandas~=1.3.4
pillow~=8.0.1
torchvision~=0.8.2
scikit-image~=0.18.1
argparse~=1.4.0
ray~=1.8.0